Remove deprecated React app files and restructure stack and queue implementations
- Deleted the entire 'docs' React app including package.json, public assets, and source files.
- Introduced new implementations for stack and queue data structures in separate files under 'linear' directory.
- Updated the main entry points for stack and queue to reference the new implementations.
